+++ newclient/src/modules/address/components/props/index.ts
@@ -18,6 +18,7 @@ export interface Tx {
     fee: FormattedNumber
     value: FormattedNumber
     status: boolean
+    contractAddress: string
 }
 
 export interface Contract {