+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 - Sort Address Tokens [#909](https://github.com/EthVM/EthVM/pull/909)
 - Add in Address Contract Details [#914](https://github.com/EthVM/EthVM/pull/914)
 
+### Devop
+
+- Contract creation [#912](https://github.com/EthVM/EthVM/pull/912)
+
 ===================================================================================
 
 ### Release v1.0.3-hotfix.2