+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Devop
 
 - Remove "To" on contract creation on Eth Transfers on Address Transfers and Transactions Table [#903](https://github.com/EthVM/EthVM/pull/903)
+- Contract creation [#912](https://github.com/EthVM/EthVM/pull/912)
 
 ===================================================================================